Switch to new backend,  update user-agent appropriately.
The big functional change is to switch from a POST to the backend to a series of GETs. This is so we can put a caching CDN in front of the backend.
Repeat requests for the same character data, even across clients, will be cached. Less work for the backend, plus some functionality will continue even if the backend is down.